Reorganize wait function
